change qtype MATCH_INDEX to integer. GF 28462 , review by dan

+++ software/caarray-ejb.jar/src/main/java/gov/nih/nci/caarray/platforms/nimblegen/NimblegenQuantitationType.java
@@ -96,7 +96,7 @@ public enum NimblegenQuantitationType implements QuantitationTypeDescriptor {
     /**
      * MATCH_INDEX.
      */
-    MATCH_INDEX(DataType.FLOAT),
+    MATCH_INDEX(DataType.INTEGER),
 
     /**
      * Perfect Match.
